Rename hotkey_handler file to hotkeys to avoid name conflict
